+++ src/Baekjoon/algorithm/graph/bfs/Q14940.java
@@ -1,4 +1,4 @@
-package Baekjoon.algorithm.graph;
+package Baekjoon.algorithm.graph.bfs;
 
 import java.io.BufferedReader;
 import java.io.IOException;
