+++ src/Baekjoon/algorithm/graph/bfs/Q18352_bfs.java
@@ -1,16 +1,17 @@
-package Baekjoon.algorithm.graph;
+package Baekjoon.algorithm.graph.bfs;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.LinkedList;
 import java.util.List;
-import java.util.PriorityQueue;
+import java.util.Queue;
 import java.util.StringTokenizer;
 
-public class Q18352_dijkstra {
-    static PriorityQueue<int[]> pq;
+public class Q18352_bfs {
+    static Queue<int[]> queue;
     static List<List<Integer>> graph;
     static boolean[] visited;
     static int[] distance;
@@ -35,18 +36,14 @@ public class Q18352_dijkstra {
             graph.get(A).add(B);
         }
 
-        pq = new PriorityQueue<>((a,b) -> a[1]-b[1]);
-        pq.add(new int[]{X,0});
+        queue = new LinkedList<>();
+        queue.add(new int[]{X,0});
 
         visited = new boolean[N+1];
+        visited[X] = true;
 
         distance = new int[N+1];
-        for(int i = 1; i <= N; i++) {
-            distance[i] = Integer.MAX_VALUE;
-        }
-        distance[X] = 0;
-
-        dijkstra();
+        bfs();
 
         List<Integer> cities = new ArrayList<>();
         for (int i = 0; i <= N; i++) {
@@ -67,23 +64,18 @@ public class Q18352_dijkstra {
 
     }
 
-    static void dijkstra() {
-        while(!pq.isEmpty()) {
-            int[] cur = pq.poll();
-            int now = cur[0];
-            int nowDis = cur[1];
+    static void bfs() {
+        while(!queue.isEmpty()) {
+            int[] cur = queue.poll();
+            int num = cur[0];
+            int dis = cur[1];
 
-            if (visited[now]) {
-                continue;
-            }
-            visited[now] = true;
+            distance[num] = dis;
 
-            for(int next: graph.get(now)){
-                int newDis = nowDis + 1;
-
-                if(distance[next] > newDis){
-                    distance[next] = newDis;
-                    pq.add(new int[]{next,newDis});
+            for(int next: graph.get(num)){
+                if(!visited[next]){
+                    queue.add(new int[]{next,dis+1});
+                    visited[next] = true;
                 }
             }
         }
